+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ### Changelog
-
+- 1.0.2: Feature: Automate "add group progress" cron job
+- 1.0.1: Fix: Show validation error message when the form is submitted
 - 1.0.0: Initial release